fix: שינוי עמודות user ID מ-Integer ל-BigInteger
מזהי Telegram יכולים לחרוג מטווח int32 (2,147,483,647).
שינוי זה מאפשר תמיכה בערכים גדולים יותר.

שינויים:
- users.id: Integer -> BigInteger
- deliveries.sender_id, courier_id: Integer -> BigInteger
- courier_wallets.courier_id: Integer -> BigInteger
- conversation_sessions.user_id: Integer -> BigInteger
- wallet_ledger.courier_id: Integer -> BigInteger
- עדכון conftest.py לתמיכה ב-SQLite בבדיקות

הערה: נדרש להריץ מיגרציה בפרודקשן:
ALTER TABLE users ALTER COLUMN id TYPE BIGINT;
ALTER TABLE deliveries ALTER COLUMN sender_id TYPE BIGINT;
ALTER TABLE deliveries ALTER COLUMN courier_id TYPE BIGINT;
ALTER TABLE courier_wallets ALTER COLUMN courier_id TYPE BIGINT;
ALTER TABLE conversation_sessions ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE wallet_ledger ALTER COLUMN courier_id TYPE BIGINT;

+++ app/db/models/delivery.py
@@ -4,7 +4,7 @@ Delivery Model - Shipment Records
 import enum
 import secrets
 from datetime import datetime
-from sqlalchemy import Column, Integer, String, DateTime, Enum as SQLEnum, Float, ForeignKey, Text
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Enum as SQLEnum, Float, ForeignKey, Text
 from sqlalchemy.orm import relationship
 
 from app.db.database import Base
@@ -33,7 +33,7 @@ class Delivery(Base):
     token = Column(String(32), unique=True, nullable=False, default=generate_secure_token, index=True)
 
     # Sender info
-    sender_id = Column(Integer, ForeignKey("users.id"), nullable=False)
+    sender_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
 
     # Pickup details
     pickup_address = Column(String(500), nullable=False)
@@ -49,7 +49,7 @@ class Delivery(Base):
 
     # Delivery info
     status = Column(SQLEnum(DeliveryStatus), default=DeliveryStatus.OPEN, index=True)
-    courier_id = Column(Integer, ForeignKey("users.id"), nullable=True)
+    courier_id = Column(BigInteger, ForeignKey("users.id"), nullable=True)
     fee = Column(Float, default=10.0)
 
     # Timestamps
